create a module with custom exceptions and rewrite CW_1
